phase_5_ui/backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
sentence-transformers
numpy
termcolor
